Update tutorial-rmd.Rmd
Clarify that `plotly` needs to be loaded first for ggplot2 to use a `SharedData` object. I tried the code in the tutorial but ran into an error since I had not realized this. I don't fully understand it, but `plotly` masks `ggplot2::last_plot` and `graphics::layout`, at least one of which is needed for ggplot2 to use a `SharedData` object.